requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1

pip~=20.2.3
py~=1.10.0
colorama~=0.4.4
setuptools~=54.0.0
future~=0.18.2
requests~=2.25.1
marshmallow~=3.10.0
PyJWT~=2.0.1
SQLAlchemy~=1.3.22
bcrypt~=3.2.0
psycopg2~=2.8.6